refactor: further decomposition
