+++ CMakeLists.txt
@@ -226,7 +226,10 @@ if (ENABLE_TESTING)
 		tests/src/unit/formatter/json
 		tests/src/unit/formatter/token
 		tests/src/unit/sink/asynchronous
-		tests/src/unit/sink/file
+		tests/src/unit/sink/file.cpp
+		tests/src/unit/sink/file/flusher/bytecount.cpp
+		tests/src/unit/sink/file/flusher/repeat.cpp
+		tests/src/unit/sink/file/stream.cpp
 		tests/src/unit/sink/null
 		tests/src/unit/sink/syslog
 		tests/src/unit/sink/tcp
